Split the DirectX12 backend into multiple files so we can break the abstraction where needed.
