+++ lib/API/Util.cpp
@@ -1,5 +1,4 @@
-#include "Util.h"
-
+#include "API/Util.h"
 #include "API/Texture.h"
 
 llvm::Error offloadtest::findAndValidateRenderPassTextureSize(
